redesigned usart handling
The only user visible setting is now the baud rate and the choice between interrupt or
polled operation. All other settings are now hidden behind the scenes. There seems to be
a bug in the serial receive interrupt routine. Using a 2byte larger input buffer solves
the problem for now only.

git-svn-id: svn://svn.code.sf.net/p/amforth/code/trunk@1333 fd1c0b69-5d68-47f5-910c-86182ebb6ecd